fix up test/dist stuff
This breaks away from varnish's example vmod build
process a bit, but the upside is now "make distcheck"
works, and it's possible to build the vmod without
first building the varnish source tree (assuming
there's an installed copy of varnish which matches
that source tree...).  It remains to be seen if
that's going to be enough to sanely build a debian
package by downloading matching varnish source.

Change-Id: I913cd254ac8c3dd860f15e291ab09e3e93773f28